requirements.txt
requests==2.32.4
pytz==2025.2
PyYAML==6.0.2
feedparser>=6.0.0
lxml>=4.9.0
python-dateutil>=2.8.0